Update
Decided To Just Add Support For Higher Numbers Without Them Having To Be Generated Off-Hand; Maybe I Will Still Add Generating The Lists For Higher Numbers As An Option In The Future, But It Isn't Really Necessary, As It Is Extremely Unlikely To Ever Need Anything Higher. I will hopefully now start working on negative numbers and imaginary numbers as well; you should expect integration with those in hopefully under a month, but we'll see how it goes and how much trouble it gives me. On one last side note, the reason I'm not pickling the lists or just storing them in a separate file is because I want one, compact script-only one file-to make it look much easier and more streamlined for the inexperienced user to run.