Update
Added much more support and greatly advanced the program to successfully complete much farther down the "quadratic process," and handle many more and different situations. Of course, this is a rough update, as with much progress so soon, comes many bugs. So, while this program is basically functional, expect many more updates soon, to fix the many known and unknown bugs haunting all over the code.