Upate
Added a function for finding the greatest common factor between two numbers, as to not rewrite the same code over and over and over. In addition to drying up the code, a few bugs were squashed and a check was added to stop the code if the a-term of the polynomial is a zero.